refactor: step-2, 리팩토링
- dto 위치 변경
  - contoller 하위가 아닌 동일 레벨로 변경
