+++ src/main/java/subway/line/service/LineService.java
@@ -3,9 +3,9 @@ package subway.line.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.line.controller.dto.LineModifyRequest;
-import subway.line.controller.dto.LineSaveRequest;
-import subway.line.controller.dto.LineResponse;
+import subway.line.dto.LineModifyRequest;
+import subway.line.dto.LineSaveRequest;
+import subway.line.dto.LineResponse;
 import subway.line.domain.Line;
 import subway.line.exception.LineNotFoundException;
 import subway.line.repository.LineCommandRepository;
