+++ src/main/java/subway/station/controller/StationController.java
@@ -2,8 +2,8 @@ package subway.station.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.station.controller.dto.StationRequest;
-import subway.station.controller.dto.StationResponse;
+import subway.station.dto.StationRequest;
+import subway.station.dto.StationResponse;
 import subway.station.service.StationService;
 
 import java.net.URI;
